app/requirements.txt
fastapi>=0.103.1
uvicorn>=0.22.0
tortoise-orm>=0.20.0
pydantic>=2.5.3
asyncpg>=0.27.0
python-multipart>=0.0.6
aiosqlite == 0.21.0
